requirements.txt
torch
xformers
mmcv==1.7.0
git+https://github.com/huggingface/diffusers
timm==0.6.12
accelerate
tensorboard
tensorboardX
transformers
sentencepiece~=0.1.99
protobuf==3.20.2
yapf==0.40.1
peft==0.6.2
ftfy
beautifulsoup4
gradio==4.44.1
gradio_imageslider==0.0.18
opencv-python
bs4
einops
optimum
came-pytorch
wandb
scipy
roma
kornia
numpy
tyro
dearpygui
imageio
imageio-ffmpeg
lpips
matplotlib
packaging
Pillow
pygltflib
rembg[gpu,cli]
rich
safetensors
scikit-image
scikit-learn
tqdm
trimesh
kiui >= 0.2.3
xatlas
plyfile